use new async fixtures for metadata service tests as well.
