+++ services/metadata_service/tests/integration_tests/flow_test.py
@@ -1,27 +1,11 @@
 from .utils import (
-    init_app, init_db, clean_db,
+    cli, db,
     assert_api_get_response, assert_api_post_response, compare_partial,
     add_flow
 )
 import pytest
-import json
+
 pytestmark = [pytest.mark.integration_tests]
-
-# Fixtures begin
-
-
-@pytest.fixture
-def cli(loop, aiohttp_client):
-    return init_app(loop, aiohttp_client)
-
-
-@pytest.fixture
-async def db(cli):
-    async_db = await init_db(cli)
-    yield async_db
-    await clean_db(async_db)
-
-# Fixtures end
 
 
 async def test_flows_post(cli, db):
