+++ services/metadata_service/tests/integration_tests/metadata_test.py
@@ -1,28 +1,13 @@
 from .utils import (
-    init_app, init_db, clean_db,
+    cli, db,
     assert_api_get_response, assert_api_post_response, compare_partial,
     add_flow, add_run, add_step,
     add_task, add_metadata
 )
 import pytest
-import json
+
 pytestmark = [pytest.mark.integration_tests]
 
-# Fixtures begin
-
-
-@pytest.fixture
-def cli(loop, aiohttp_client):
-    return init_app(loop, aiohttp_client)
-
-
-@pytest.fixture
-async def db(cli):
-    async_db = await init_db(cli)
-    yield async_db
-    await clean_db(async_db)
-
-# Fixtures end
 
 # Shared metadata test data
 METADATA_A = {
